refactor: DetailService와 TaskService에 각자 있던 updatePassedRate을 WorkService를 만들어서 하나로 통합
