+++ src/main/java/com/ideality/coreflow/project/command/application/service/DetailService.java
@@ -19,6 +19,4 @@ public interface DetailService {
     void completeDetail(Long workId);
 
     void deleteDetail(Long workId);
-
-    Double updateDetailPassedRate(Long detailId);
 }
