+++ src/main/java/com/ideality/coreflow/project/command/application/service/WorkService.java
@@ -1,0 +1,5 @@
+package com.ideality.coreflow.project.command.application.service;
+
+public interface WorkService {
+    Double updatePassedRate(Long workId);
+}
