+++ src/main/java/com/ideality/coreflow/project/command/application/service/impl/TaskServiceImpl.java
@@ -26,7 +26,6 @@ import static com.ideality.coreflow.common.exception.ErrorCode.*;
 public class TaskServiceImpl implements TaskService {
 
     private final TaskRepository taskRepository;
-    private final HolidayQueryService holidayQueryService;
 
     @Override
     @Transactional
@@ -117,26 +116,5 @@ public class TaskServiceImpl implements TaskService {
         if (!taskRepository.existsById(taskId)) {
             throw new BaseException(TASK_NOT_FOUND);
         }
-    }
-
-    @Override
-    public Double updateTaskPassedRate(Long taskId) {
-        Work work = taskRepository.findById(taskId).orElseThrow(() -> new BaseException(TASK_NOT_FOUND));
-        LocalDate now = LocalDate.now();
-        LocalDate endBase = work.getEndBase();
-        LocalDate startBase = work.getStartBase();
-        LocalDate startReal = work.getStartReal();
-
-        Long totalDuration = ChronoUnit.DAYS.between(startBase, endBase)+1
-                            -holidayQueryService.countHolidaysBetween(startBase, endBase);
-        Long passedDates = ChronoUnit.DAYS.between(startReal, now)+1
-                            -holidayQueryService.countHolidaysBetween(startReal, now);
-        Double passedRate = (double) passedDates / totalDuration * 100;
-        passedRate = passedRate>100?100:Math.round(passedRate*100)/100.0;
-        work.setPassedRate(passedRate);
-        taskRepository.saveAndFlush(work);
-        return work.getPassedRate();
-
-
     }
 }
