+++ src/main/java/com/ideality/coreflow/project/query/controller/WorkController.java
@@ -1,16 +1,12 @@
 package com.ideality.coreflow.project.query.controller;
 
-import com.ideality.coreflow.calendar.query.dto.RequestUserDTO;
 import com.ideality.coreflow.common.response.APIResponse;
-import com.ideality.coreflow.project.query.dto.DeptWorkDTO;
 import com.ideality.coreflow.project.query.dto.DetailDTO;
 import com.ideality.coreflow.project.query.dto.WorkDetailDTO;
-import com.ideality.coreflow.project.query.service.WorkService;
+import com.ideality.coreflow.project.query.service.WorkQueryService;
 import lombok.RequiredArgsConstructor;
 
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -22,7 +18,7 @@ import java.util.List;
 @RequestMapping("/api/work")
 public class WorkController {
 
-    private final WorkService workService;
+    private final WorkQueryService workService;
 
     // 태스크 내의 세부일정 이름 목록 조회
     @GetMapping("/detail/nameList")
