+++ src/main/java/com/ideality/coreflow/project/query/service/facade/ProjectQueryFacadeService.java
@@ -11,7 +11,7 @@ import com.ideality.coreflow.project.query.service.ProjectQueryService;
 import com.ideality.coreflow.project.query.service.RelationQueryService;
 import com.ideality.coreflow.project.query.service.TaskQueryService;
 import com.ideality.coreflow.project.query.service.WorkDeptQueryService;
-import com.ideality.coreflow.project.query.service.WorkService;
+import com.ideality.coreflow.project.query.service.WorkQueryService;
 import com.ideality.coreflow.user.query.service.UserQueryService;
 
 import lombok.RequiredArgsConstructor;
@@ -29,7 +29,7 @@ public class ProjectQueryFacadeService {
     private final UserQueryService userQueryService;
     private final DeptQueryService deptQueryService;
     private final RelationQueryService relationQueryService;
-    private final WorkService workService;
+    private final WorkQueryService workService;
     private final WorkDeptQueryService workDeptQueryService;
     private final ProjectQueryService projectQueryService;
