+++ src/main/java/com/ideality/coreflow/project/query/service/impl/WorkQueryServiceImpl.java
@@ -5,7 +5,7 @@ import com.ideality.coreflow.project.query.dto.DetailDTO;
 import com.ideality.coreflow.project.query.dto.ParticipantDTO;
 import com.ideality.coreflow.project.query.dto.WorkDetailDTO;
 import com.ideality.coreflow.project.query.mapper.WorkMapper;
-import com.ideality.coreflow.project.query.service.WorkService;
+import com.ideality.coreflow.project.query.service.WorkQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,7 +14,7 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class WorkServiceImpl implements WorkService {
+public class WorkQueryServiceImpl implements WorkQueryService {
 
     private final WorkMapper workMapper;
